refactor: encapsulate lotto number logic in LottoNumbers class and update Game class accordingly
