+++ src/main/java/lotto/LottoTicketMachine.java
@@ -7,15 +7,15 @@ import java.util.stream.IntStream;
 
 public class LottoTicketMachine {
 
-    private final List<Integer> availableNumbers = IntStream.rangeClosed(1, 45).boxed().collect(Collectors.toUnmodifiableList());;
+    private final List<Integer> availableNumbers = IntStream.rangeClosed(1, 45).boxed().collect(Collectors.toUnmodifiableList());
     private final RandomPicker randomPicker;
 
     public LottoTicketMachine(Random random) {
         this.randomPicker = new RandomPicker(random);
     }
 
-    public LottoTicket issue() {
-        var lottoNumber = randomPicker.pickNumbers(Game.DIGIT, availableNumbers);
-        return new LottoTicket(lottoNumber);
+    public Game issue() {
+        var lottoNumber = randomPicker.pickNumbers(LottoNumbers.DIGIT, availableNumbers);
+        return new Game(lottoNumber);
     }
 }
